requirements.txt
Flask-Session==0.5.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.3

